customer-sentiment-analyzer/requirements.txt
streamlit>=1.28.0
anthropic>=0.18.0
pandas>=2.0.0
openpyxl>=3.1.0
plotly>=5.18.0
jinja2>=3.1.0
python-dotenv>=1.0.0
tenacity>=8.2.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
